refactor: move environment table into new environment folder
Will add more environment related datatables
