+++ EMS/core-bundle/src/Resources/config/datatable.xml
@@ -29,6 +29,11 @@
             <argument>%ems_core.template_namespace%</argument>
             <tag name="emsco.datatable" />
         </service>
+        <service id="emsco.data_table.environment" class="EMS\CoreBundle\DataTable\Type\Environment\EnvironmentDataTableType">
+            <argument type="service" id="emsco.core.environment.managed_entity"/>
+            <argument>%ems_core.template_namespace%</argument>
+            <tag name="emsco.datatable" />
+        </service>
         <service id="emsco.data_table.mapping.analyzer" class="EMS\CoreBundle\DataTable\Type\Mapping\AnalyzerDataTableType">
             <argument type="service" id="emsco.helper.analyzer"/>
             <tag name="emsco.datatable" />
@@ -92,11 +97,6 @@
         </service>
         <service id="emsco.data_table.channel" class="EMS\CoreBundle\DataTable\Type\ChannelDataTableType">
             <argument type="service" id="ems.service.channel"/>
-            <tag name="emsco.datatable" />
-        </service>
-        <service id="emsco.data_table.environment" class="EMS\CoreBundle\DataTable\Type\EnvironmentDataTableType">
-            <argument type="service" id="emsco.core.environment.managed_entity"/>
-            <argument>%ems_core.template_namespace%</argument>
             <tag name="emsco.datatable" />
         </service>
         <service id="emsco.data_table.wysiwygProfile" class="EMS\CoreBundle\DataTable\Type\Wysiwyg\WysiwygProfileDataTableType">
